实体类变量使用包装类 修正了sql缺少的表单

+++ src/main/java/com/softwarelearn/lite_mall_backend/pojo/mall/VisitNumInfo.java
@@ -13,6 +13,6 @@ import java.io.Serializable;
 @Slf4j
 @Data
 public class VisitNumInfo implements Serializable {
-    private int id;
-    private int visitNum;
+    private Long id;
+    private Integer visitNum;
 }
